(RE-3884) Setup C++ component builds in Vanagon for SLES
This commit adds boost, cmake, gcc, yaml-cpp and make (for SLES 10) to
this vanagon project. The project is designed to build out our C++ stack
and tools for each target.

Items of note are that there are multiple projects in the project folder
here meaning there are a few different build targets options. (This
isn't just one package). There are packages for

    * GCC
    * cmake
    * boost
    * yaml-cpp
    * make (only required for SELS 10)

There are a few minor changes from using the existing pl-build-tools
that were built from pl-dependencies. Since vanagon doesn't do
sub-projects, we have single packages with static and dynamic
libraries.