(maint) Remove files with spaces
Debian has a grudge against me, and totally bombs out when you try to
package files that have spaces in their names. Well, not debian per se,
but debhelper. Since the particular files that are causing issues here
are needed when we run the tests against cmake, we need them around. But
we can delete them once those tests are run, and not worry about it.
Thanks debhelper. You're just so helpful.