(RE-4378) Fix broken SLES 10 GCC build
During a series of moves over from pl-dependencies, we messed up an if
statement and accidentally required packages on SLES 10 that are not
available. This commit fixes that, and SLES 10 should be able to build
again.

Since we're not running the GCC test suite anywhere, we are just
removing the BR on those packages. I've left a comment in the source
code so that in the event we want to enable the suite in the future,
we'll know what packages to bring in on what platforms.