(RE-4240) Update Boost to 1.57
This commit updates boost to 1.57.

In building boost at 1.57 they removed the v2 directory, so now we take
that out. Also, we're just dropping the user-config.jam file in $HOME
since that's where it looks by default.

The configure section of the boost config was pretty difficult to read
and understand. This cleans it up a bit and makes each entry in the
array have its own line.