(maint) Collect all project settings in one file
Previously the same settings were repeated in each new project, which
leads to extra work when changing a setting, and drift between projects
if a setting is changed in one project and not others. This commit tries
to avoid that problem by adding a single file with all of the common
settings that can be sourced in each of the projects.