(RE-2452) Add toolchain files as sources
This is a commit that removes the wget/curl usage to grab toolchain
files and instead copies in specific toolchain files. Right now this
works for Linux and AIX. Included in this commit are the origianl
toolchain files for Solaris, but those will likely need to be modified
as the vanagon progress commences.