(RE-2371) Build GCC for AIX 5.3
Because Geoff was able to do some big Ramdisk-fu, we were finally able
to build gcc for AIX 5.3. We used a ramdisk for /root, /tmp and I
symlinked in /var/tmp.

I also used a tarball of gcc464 that I built earlier outside of vanagon
to bootstrap it.
http://pl-build-tools.delivery.puppetlabs.net/aix/5.3/ppc/gcc464-aix53.tar.gz

The only real changes needed here are to adjust the target string that
gcc uses so we are explicit about aix-5.3 vs the 6.1 binary
compatibility.