(maint) AIX setup command can't require a tty
On AIX if one of the commands we send over ssh requires a tty, (or at
least things like echo) the command just hangs forever. This doesn't
really work for automated builds. This commit just places a # in front
of the command since that just works (and that's what it was before Haus
told me to change it)