(RE-5093) Add EL4 section to linux toolchain for cmake
On el4, the linker is not very helpful. It doesn't, for example, link
against ld-linux.so by default. This makes finding some functions like
tls_get_addr not work at all on el4. This commit adds some logic to the
linux toolchain to explicitly add ld-linux.so to the link calls so that
these symbols can be correctly resolved.