(RE-4563) Update project settings to include basedir and platform_triple
When building binutils for solaris, we will be passing --target, which
will add a suffix to the supplied prefix. This means we will want a
directory that is just /opt/pl-build-tools for that purpose. We also
want bindir, libdir and friends to include the triple in their path for
solaris, so we do want prefix to include it. This commit adds a new
setting called basedir to allow that. It also includes a new setting
called platform_triple which can be used in binutils and gcc builds as
the target.