(RE-5145) Add solaris-10-sparc-sysroot component for gcc
To build a functional cross compiler, the system headers and libraries
from the target system are required. This commit adds a component to
deploy these headers and libraries for use in the gcc build. They will
be included in the sparc gcc package as well. This also updates the
binutils dependency to pull in the correct architecture.