(RE-5146) Update solaris 10 toolchain files
Because we are cross compiling for solaris 10, we will be using the i386
package to build both sparc and i386 packages (we just change which
toolchain file we're using). Because vanagon does not currently support
multiple sources in a single component, this commit adds a tarball of
the current solaris 10 toolchains and updates the component to install
both toolchains. It also moves a chmod on the toolchain file from cmake
into the toolchain component.