(RE-5146) Refactor toolchain installation
Previously the install created directories and copied files, when there
is a dsl method that does all of that. This commit moves the component
to use that method instead and also adds a chown to ensure that the
toolchains are added to the package on solaris (it will quietly not add
them to the package if they are owned by an invalid user).