(RE-5245) Apply patch to use g++ as the linker for solaris
Previously, due to deficiencies in solaris' linking, cmake decided to
always use gcc to link, even for c++ code. This has the downside of not
working without explicitly linking in libstdc++, which is tedious. This
changed recently in cmake's git repo, so this PR applies that patch to
ensure that cmake uses g++ when appropriate.