(RE-5237) Build most of boost
puppet-access wanted access to the iostreams component of boost;
however, we were not building that. In a discussion about boost options,
we realized were building less than half of the boost subcomponents, and
that many other boost components could prove to be useful in the future.

This commit builds out all boost components with the exeception of
python bindings, icu (unicode stuff), and mpi. Each of those brings
along new dependencies and should probably have a solid use-case
before requesting them.

In short we're now building:

    * atomic
    * container
    * context
    * coroutine
    * exception
    * graph
    * graph_parallel
    * iostreams
    * math
    * serialization
    * signals
    * test
    * timer
    * wave

In addition to:

    * chrono
    * date_time
    * filesystem
    * locale
    * log
    * program_options
    * random
    * regex
    * system
    * thread

This also bumps the version of boost to 1.58.

Note: this requires https://github.com/puppetlabs/vanagon/pull/131 to be
merged for all platforms to go green.