(maint) Compile flags and user-jam updates
Some modules require -std=c99 to be passed in order to compile cleanly,
so this commit adds that to cflags and also breaks up cflags and
cxxflags into distinct variables. As both are safe to apply on any
platform using our gcc, they are lifted out of the conditional. This
commit also adds single quotes around the user-config.jam line to ensure
that it is correctly parsed during builds.