(RE-5290) Use shared libraries on solaris
It turns out that cross-compiling just isn't going to work out well
with static linking. Also, it's a giant pain in solaris and we probably
want to start walking toward dynamic linking as much as possible given
the woes with el4, sol, aix. This commit makes such a change by removing
the special solaris conditional for gcc, so libgcc.so and libstdc++.so
will now be generated.