(maint) Update boost on solaris to build 1.58.0
Boost 1.58.0 wouldn't build on solaris for 2 main reasons. First, the
build script had bashisms that wouldn't work in the default shell on
solaris. Second, the sparc version was set incorrectly for sparc builds.
This commit works around the first problem by applying a patch to boost
and works around the second problem by specifying the instruction set to
target for sparc builds.