(RE-5294) Set LDFLAGS_FOR_TARGET for linux gcc builds
When building with --disable-shared, libssp and libstdc++ won't need to
find libgcc_s, because they won't ever be linked dynamically. This is
changing as our toolchain moves away from static linking, as it causes
odd linking bugs. Because of this, we must now begin setting an rpath
for gcc's libs. This commit copies the work from the solaris conditional
into the linux conditional to achieve this.