(maint) setup GCC build_requires on binutils, m4, autoconf for AIX
This commit updates the location for gcc to build autoconf, m4 and
binutils. These are built using our GCC because of odd issues,
specifically on AIX 7.1.

Also note that if you have system gcc, libgcc installed on the AIX
system at the time you attempt to build autoconf or binutils, it
probably won't work. So, clean up your work when you're done.

This also remove the g++ BR from the linux systems, because it wasn't
real for m4, autoconf or binutils.