(RE-5240) Update binutils and gcc for solaris 11 sparc
Solaris 11 packages are more stringent with architecture than solaris
10, which makes installing packages for a cross-compile more
challenging. To get around this, we add the architecture to the project
name, so that we can have two packages in the repository for binutils
(one for i386, one for sparc). Noarch must also be set for these to
install correctly, even though it is a total lie.