(RE-5579) Update solaris 10 platform configs for older template
In order to rebuild on an older solaris template, a few tweaks are
required. First, the vcloud_name is updated to refer to the u8 template.
This template is more minimal than the newer update, so more
dependencies such as gcc are required. Ntpdate also isn't present on the
older template, but it shouldn't be needed anyway.