(RE-5648) Update and unify the cmake toolchain files
This commit goes through all of our toolchain files and updates the
CFLAGS, CXXFLAGS and RPATH to be consistent across platforms while
ensuring that the agent's include dir is searched for headers and libdir
is searched for libs. Additionally, we specify CMAKE_INSTALL_PREFIX so
that it will default to /opt/puppetlabs/puppet and not need to be
specified at build time in facter, pcp or pxp.