(maint) Create a toolchain package
This work is a little silly, but makes my life easier at the moment.
CMake is proving challenging to build on windows. Since we know the
currently built cmake nuget package works, the only thing we need to
make available is the toolchain file. As a result, an interim step here
is to just make the cmake toolchain file available in its own package so
that other projects can take advanage of it.