(maint) Remove the default chocolatey repo
We want to ensure we are only installing packages from internal feeds.
This commit removes the only external source that chocolatey makes
available by default. This is important because we have yet to make our
install command smart. We can't pass in the version or the source to
install from, so this is our best way forward at the moment.