(maint) Enable builds not on Puppet VPN
Prior to this commit, all referenced URIs were on our internal systems.
This means that building any of our toolchain is very difficult. This
commit flips components to using the upstream URL by default, however it
then uses the vanagon rewrite engine in pl-build-tools.rb to rewrite the
download URI to buildsources.delivery.puppetlabs.net.

If you're building and not on the Puppet VPN, you'd need to comment out
the rewrite line in pl-build-tools.rb