(maint) Sysroot location update
In order to enable the vanagon rewrite engine, all rewrites have to
target the same area. This means all content needs to be on
buildsources rather than having some items pulled from pl-build-tools
and some from buildsources. The net result here is that all sysroots
need to be on buildsources and ideally, nowhere else. The sysroots have
been relocated to buildsources and the component files updated
accordingly.