(RE-7315) Remove static build of gcc
We were building gcc with '--disable-shared' on some platforms. We
thought we were doing that on most platforms, but due to a logic issue,
that was not the case. After more analysis and discussion, we've elected
to always build the shared libraries for libstdc++ and libgcc and we'll
copy them into the projects we need via a runtime component (ala
puppet-agent today). This will allow for consistency cross-platform and
allow us to write tests to ensure we're linked appropriately.