(maint) Bump release versions for build-tools used by solaris 10
A while back we changed the template used for solaris 10 builds to
an older version (u8) to improve compatibility of our solaris packages.
We need to rebuild everything on the u8 template for libc compatibility,
as trying to use the previously compiled packages results in build
errors SUNW_1.22.6 missing in libc.so.1.