(BUILD-68) Fix gettext 0.19.8 build on EL5
On EL5, there is a bug in the emacs scripts that gettext tries to
install when it is built. These scripts are used to enable a special
emacs mode for translators editing PO files. Since we don't need that
here at all, this commit disables the emacs addons on all platforms.