(arch) Iinitial support for AArch64
A first attempt at providing a vanogon build tools for CentOS7 on
AArch64 architecture. This patch assumes that the build will be
performed on AArch64 hardware. It also assumes (probably incorrectly)
that a vmpooler_template is available for aarch64.

I have been in contact with @ahpook from the puppet team and he
may be able to provide additional help in fixing this patch with
respect to puppetlabs internal build setup.

Signed-off-by: Richard Henwood <richard.henwood@arm.com>